@@ -1,8 +1,8 @@
 package com.pokemonreview.api.models;
 
+import jakarta.persistence.*;
 import lombok.Data;
 
-import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
 
